https://github.com/adafruit/Adafruit_IS31FL3731/pull/11/files

+++ Adafruit_IS31FL3731.cpp
@@ -272,7 +272,7 @@ uint8_t  Adafruit_IS31FL3731::readRegister8(uint8_t bank, uint8_t reg) {
  Wire.write((byte)reg);
  Wire.endTransmission();
 
- Wire.requestFrom(_i2caddr, (byte)1);
+ Wire.requestFrom(_i2caddr, (size_t)1);
  x = Wire.read();
 
 // Serial.print("$"); Serial.print(reg, HEX);
